Fixed begin(cspin) routine With the old setup begin(cspin) did not correctly set *ldac* to false. It was overridden by calling begin(cspin,0).
I switched it around a bit, because I thought that setting ldacpin to
output (even if it was 0) everytime did not make much sense.

So instead of begin(cspin) calling begin(cspin,0) it is now the other
way around. begin(cspin,latchpin) is only called externally and in
returns calls the begin(cspin) and sets up latching afterwards.